Added a preview callback.
The preview callback receives an Image object, consistent with the Camera2 API.
Packed/interleaved Camera1 image formats are stored on a single Plane in the Image.
The preferred format for the preview can be specified. If available, that format will
be used and if not, will fallback to whatever the default format is.